refactor: ExceptionMessage 클래스 의존성 줄이는 방향으로 분리
