+++ src/main/java/lotto/view/exception/ExceptionMessage.java
@@ -1,0 +1,10 @@
+package lotto.view.exception;
+
+public class ExceptionMessage {
+
+    private ExceptionMessage() {
+    }
+
+    public static final String INVALID_LOTTO_NUMBER_FORMAT =
+            "번호는 숫자여야 합니다.";
+}
